refactor(cli): extract auth flows into per-provider modules
Move provider-specific auth logic from the monolithic app.rs into
crates/acrawl-cli/src/auth/ with per-provider files:

- auth/anthropic.rs: Anthropic OAuth PKCE + API key flows
- auth/openai.rs: OpenAI OAuth PKCE + API key flows
- auth/custom.rs: base URL + optional API key flow
- auth/mod.rs: dispatch, credential persistence, provider selection

Adding a new provider's auth now means adding one file to auth/
instead of growing match arms in app.rs.
